// src/app/model/kecamatanModel.ts
export interface Kecamatan {
  no: any;
  districtId: any;
  districtName: any;
  cityName: any;
  provinceName: any;
  countryNameIdn: any;
}

export interface DataKecamatan {
  no: any;
  districtName: any;
  districtId: any;
  cityName: any;
  cityId: any;
  provinceName: any;
  provinceId: any;
  countryNameIdn: any;
  countryId: any;
}